refactor(server): extract game manager
- move game start/restart, snapshot broadcasting, and game cleanup
  behind a dedicated manager while preserving existing gameplay
