+++ src/server.ts
@@ -45,6 +45,7 @@ import { ConnectionManager } from './server/ConnectionManager';
 import { DisconnectHandler } from './server/DisconnectHandler';
 import { LobbyManager } from './server/LobbyManager';
 import { PersistenceService } from './server/PersistenceService';
+import { GameManager } from './server/GameManager';
 import {
   PlayerInfo,
   PlayerInLobby,
@@ -326,14 +327,14 @@ if (process.env.NODE_ENV !== 'production') {
         // Find and clear games the user is in
         const lobbyId = lobbyManager.getLobbyIdForPlayer(userId);
         if (lobbyId && gameLoopsByLobbyId.has(lobbyId)) {
-          clearActiveGameArtifacts(lobbyId);
+          gameManager.clearActiveGameArtifacts(lobbyId);
           gamesCleared++;
           logger.info(`[TEST] Cleared game for user ${userId} in lobby ${lobbyId}`);
         }
       } else {
         // Clear all games
         gameLoopsByLobbyId.forEach((gameLoop, lobbyId) => {
-          clearActiveGameArtifacts(lobbyId);
+          gameManager.clearActiveGameArtifacts(lobbyId);
           gamesCleared++;
         });
         logger.info(`[TEST] Cleared all games: ${gamesCleared}`);
@@ -468,7 +469,6 @@ clearPlayerDisconnectTimerWrapper.fn = (playerId: string) =>
 // Create PersistenceService instance
 const persistenceService = new PersistenceService(logger);
 
-
 /**
  * Clean up bots that are not part of any active game
  */
@@ -479,30 +479,24 @@ function cleanupInactiveBots(): void {
   );
 }
 
-function clearActiveGameArtifacts(lobbyId: string): Lobby | undefined {
-  const loop = gameLoopsByLobbyId.get(lobbyId);
-  if (loop) {
-    loop.cancel();
-    loop.removeAllListeners();
-    gameLoopsByLobbyId.delete(lobbyId);
-  }
-
-  mostRecentGameSnapshotByLobbyId.delete(lobbyId);
-  currentRoundGameEventsByLobbyId.delete(lobbyId);
-  roundStartSnapshotByLobbyId.delete(lobbyId);
-  supabaseGameIdByLobbyId.delete(lobbyId);
-  gameIdByLobbyId.delete(lobbyId);
-  cleanupBots(lobbyId);
-
-  const lobby = lobbyManager.getLobby(lobbyId);
-  if (lobby) {
-    lobby.disconnectedPlayerIds.forEach(playerId =>
-      disconnectHandler.clearPlayerDisconnectTimer(playerId)
-    );
-  }
-
-  return lobby;
-}
+// Create GameManager instance
+const gameManager = new GameManager({
+  io,
+  connectionManager,
+  lobbyManager,
+  disconnectHandler,
+  persistenceService,
+  gameLoopsByLobbyId,
+  mostRecentGameSnapshotByLobbyId,
+  currentRoundGameEventsByLobbyId,
+  roundStartSnapshotByLobbyId,
+  supabaseGameIdByLobbyId,
+  currentGameBotIdsByLobbyId,
+  gameIdByLobbyId,
+  cleanupBots,
+  emitConnectedPlayers,
+  SUPABASE_AUTH_ENABLED,
+});
 
 // Cleanup timer is now managed by LobbyManager
 
@@ -528,7 +522,7 @@ registerHttpApi(app, {
       `Player ${playerId} left lobby ${lobbyId} via HTTP API - cleared in-memory state`
     );
   },
-  onStartLobbyGame: (lobbyId, hostId) => startLobbyGameForHost(lobbyId, hostId),
+  onStartLobbyGame: (lobbyId, hostId) => gameManager.startLobbyGameForHost(lobbyId, hostId),
 });
 
 async function refreshLobbyFromSupabase(
@@ -711,7 +705,7 @@ io.on('connection', socket => {
         'lobbyId:',
         data?.lobbyId
       );
-      handleStartLobbyGame(socket, data, callback).catch(error => {
+      gameManager.handleStartLobbyGame(socket, data, callback).catch(error => {
         logger.error('Error starting lobby game:', error);
         if (callback) callback({ error: 'Failed to start game' });
         socket.emit('error', { message: 'Failed to start game' });
@@ -721,7 +715,7 @@ io.on('connection', socket => {
 
   socket.on('restartGame', () => {
     logger.info('Received restartGame event from socket:', socket.id);
-    handleRestartGame(socket).catch(error => {
+    gameManager.handleRestartGame(socket).catch(error => {
       logger.error('Error restarting game:', error);
       socket.emit('error', { message: 'Failed to restart game' });
     });
@@ -775,304 +769,7 @@ io.on('connection', socket => {
 });
 
 // Lobby handlers moved to LobbyManager
-
-async function startLobbyGameForHost(
-  lobbyId: string,
-  hostId: string
-): Promise<{ lobby: LobbyPayload; gameId: string }> {
-  if (gameLoopsByLobbyId.has(lobbyId)) {
-    throw new Error('GAME_IN_PROGRESS');
-  }
-
-  const startedLobby = await startLobby({ lobbyId, hostId });
-  const lobby = lobbyManager.cacheLobbyFromPayload(startedLobby);
-  const newBotIds: string[] = [];
-
-  try {
-    // Clean up any existing bots before creating new ones
-    cleanupBots(lobby.id);
-
-    // Build playersInfo from lobby members (humans only, no bots yet)
-    const playersInfo: PlayerIdAndName[] = lobby.players.map(p => ({
-      id: p.playerId,
-      name: p.displayName,
-    }));
-
-    // Calculate bots needed
-    const targetCount =
-      lobby.maxPlayers ?? lobby.playerCount ?? playersInfo.length;
-    const botsNeeded = Math.max(0, targetCount - playersInfo.length);
-    logger.info(
-      `Starting lobby game: ${lobby.name} with ${playersInfo.length} humans and ${botsNeeded} bots needed`
-    );
-
-    // Create bots
-    const botNames = ['Bot Alex', 'Bot Morgan', 'Bot Jordan'];
-    for (let i = 0; i < botsNeeded; i++) {
-      const botName = botNames[i] || `Bot ${i + 1}`;
-      const botAgent = new ExhaustiveSimpleAgent();
-      const botId = `${botAgent.playerId}-${Date.now()}-${i}`;
-      // Update the agent's playerId to match the generated botId
-      botAgent.playerId = botId;
-      playersInfo.push({ id: botId, name: botName });
-      const botPlayerInfo: PlayerInfo = {
-        id: botId,
-        name: botName,
-        agent: botAgent,
-      };
-      connectionManager.setPlayer(botId, botPlayerInfo);
-      newBotIds.push(botId);
-      logger.info(`Added bot: ${botName} (ID: ${botId})`);
-    }
-
-    // Create GameLoop using players from the lobby
-    const agents: Map<string, GameAgent> = new Map();
-    // Populate human agents
-    lobby.players.forEach(p => {
-      const info = connectionManager.getPlayer(p.playerId);
-      if (info) agents.set(info.id, info.agent);
-    });
-    // Populate bot agents
-    newBotIds.forEach(id => {
-      const info = connectionManager.getPlayer(id);
-      if (info) agents.set(info.id, info.agent);
-    });
-
-    // Filter out disconnected players - only include players who have agents
-    const validPlayersInfo = playersInfo.filter(p => agents.has(p.id));
-    if (validPlayersInfo.length !== playersInfo.length) {
-      const disconnectedPlayers = playersInfo.filter(p => !agents.has(p.id));
-      logger.warn(
-        `[startLobbyGameForHost] Filtering out ${
-          disconnectedPlayers.length
-        } disconnected players: ${disconnectedPlayers
-          .map(p => p.name)
-          .join(', ')}`
-      );
-    }
-    if (validPlayersInfo.length < 2) {
-      throw new Error('Not enough connected players to start game');
-    }
-
-    // Store bot IDs for cleanup after game ends
-    currentGameBotIdsByLobbyId.set(lobby.id, newBotIds);
-
-    const gameLoop = new GameLoop(validPlayersInfo);
-    agents.forEach((agent, id) => gameLoop.addAgent(id, agent));
-    gameLoopsByLobbyId.set(lobby.id, gameLoop);
-    currentRoundGameEventsByLobbyId.set(lobby.id, []);
-    await persistenceService.createSupabaseGameForLobby(
-      lobby,
-      validPlayersInfo,
-      gameLoop,
-      supabaseGameIdByLobbyId,
-      SUPABASE_AUTH_ENABLED
-    );
-
-    // Set up gameSnapshot listener to send redacted snapshots to all clients
-    let firstSnapshotEmitted = false;
-    gameLoop.on('gameSnapshot', (newSnapshot: GameSnapshot) => {
-      mostRecentGameSnapshotByLobbyId.set(lobby.id, newSnapshot);
-      const existingEvents =
-        currentRoundGameEventsByLobbyId.get(lobby.id) || [];
-      const updatedEvents = [...existingEvents, newSnapshot.gameEvent];
-      const isStartRound =
-        newSnapshot.gameEvent.actionType === ActionType.START_ROUND;
-      const roundEvents = isStartRound
-        ? [newSnapshot.gameEvent]
-        : updatedEvents;
-      if (isStartRound) {
-        logger.debug(
-          `[Supabase] START_ROUND detected, resetting event collection for lobby ${lobby.id}`
-        );
-        roundStartSnapshotByLobbyId.set(lobby.id, newSnapshot);
-        currentRoundGameEventsByLobbyId.set(lobby.id, roundEvents);
-      } else {
-        currentRoundGameEventsByLobbyId.set(lobby.id, updatedEvents);
-        logger.debug(
-          `[Supabase] Collected event ${newSnapshot.gameEvent.actionType} (lobby ${lobby.id}, total events: ${updatedEvents.length})`
-        );
-      }
-
-      if (
-        newSnapshot.gameEvent.actionType === ActionType.READY_FOR_NEXT_ROUND ||
-        newSnapshot.gameEvent.actionType === ActionType.WIN
-      ) {
-        const eventsToPersist =
-          currentRoundGameEventsByLobbyId.get(lobby.id) || [];
-        logger.info(
-          `[Supabase] Triggering persistence for ${newSnapshot.gameEvent.actionType} (lobby ${lobby.id}, ${eventsToPersist.length} events collected)`
-        );
-      // Use void to fire-and-forget, but ensure errors are logged in persistRoundHistory
-      void persistenceService
-        .persistRoundHistory(
-          lobby.id,
-          newSnapshot,
-          supabaseGameIdByLobbyId,
-          currentRoundGameEventsByLobbyId,
-          roundStartSnapshotByLobbyId,
-          SUPABASE_AUTH_ENABLED
-        )
-        .catch(error => {
-          logger.error(
-            `[Supabase] Unhandled error in persistRoundHistory for lobby ${lobby.id}`,
-            error
-          );
-        });
-      }
-
-      // Send redacted snapshots to all players
-      sendRedactedSnapshotToAllPlayers(
-        gameLoop,
-        newSnapshot,
-        roundEvents,
-        lobby.id
-      );
-
-      // After the first snapshot, ensure all clients received it by re-emitting
-      // This helps clients that reset state after gameStartedFromLobby
-      if (!firstSnapshotEmitted) {
-        firstSnapshotEmitted = true;
-        // Small delay to ensure clients have processed gameStartedFromLobby and set up listeners
-        setTimeout(() => {
-          const latestSnapshot = mostRecentGameSnapshotByLobbyId.get(lobby.id);
-          const latestEvents =
-            currentRoundGameEventsByLobbyId.get(lobby.id) || [];
-          if (latestSnapshot === newSnapshot) {
-            logger.info(
-              'Re-emitting first game snapshot to ensure all clients received it'
-            );
-            sendRedactedSnapshotToAllPlayers(
-              gameLoop,
-              latestSnapshot,
-              latestEvents,
-              lobby.id
-            );
-          }
-        }, 100);
-      }
-    });
-
-    // Map lobby -> game
-    const gameId = gameLoop.cribbageGame.getGameState().id;
-    gameIdByLobbyId.set(lobby.id, gameId);
-
-    // Update lobby status and broadcast updates
-    lobby.status = 'in_progress';
-    lobby.finishedAt = null;
-    lobby.disconnectedPlayerIds = [];
-    io.emit('lobbyUpdated', lobby);
-
-    // Notify lobby members of the game start
-    io.emit('gameStartedFromLobby', {
-      lobbyId: lobby.id,
-      gameId,
-      players: validPlayersInfo,
-    });
-
-    // Start the game loop (this will emit snapshots as the game progresses)
-    // Don't await - let it run in the background
-    startGame(lobby.id).catch(error => {
-      logger.error('[startLobbyGameForHost] Error in game loop:', error);
-    });
-
-    return { lobby: startedLobby, gameId };
-  } catch (error) {
-    // Rollback lobby status to 'waiting' if game start fails
-    try {
-      const client = getServiceClient();
-      await client
-        .from('lobbies')
-        .update({ status: 'waiting' })
-        .eq('id', lobbyId);
-      logger.info(
-        `[startLobbyGameForHost] Rolled back lobby ${lobbyId} status to 'waiting' due to error`
-      );
-    } catch (rollbackError) {
-      logger.error(
-        `[startLobbyGameForHost] Failed to rollback lobby ${lobbyId} status:`,
-        rollbackError
-      );
-    }
-
-    // Clean up any bots that were created
-    newBotIds.forEach(botId => {
-      connectionManager.deletePlayer(botId);
-    });
-    if (newBotIds.length > 0) {
-      logger.info(
-        `[startLobbyGameForHost] Cleaned up ${newBotIds.length} bots after error`
-      );
-    }
-
-    // Clean up any partial game state that may have been created
-    gameLoopsByLobbyId.delete(lobby.id);
-    currentGameBotIdsByLobbyId.delete(lobby.id);
-    gameIdByLobbyId.delete(lobby.id);
-    currentRoundGameEventsByLobbyId.delete(lobby.id);
-    mostRecentGameSnapshotByLobbyId.delete(lobby.id);
-    roundStartSnapshotByLobbyId.delete(lobby.id);
-
-    // Re-throw the original error
-    throw error;
-  }
-}
-
-// Lobby handlers moved to LobbyManager
-
-async function handleStartLobbyGame(
-  socket: Socket,
-  data: { lobbyId: string },
-  callback?: (response: any) => void
-): Promise<void> {
-  const playerId = connectionManager.getPlayerId(socket.id);
-  if (!playerId) {
-    logger.error('Player ID not found for socket:', socket.id);
-    const errorMsg = 'Not logged in';
-    socket.emit('error', { message: errorMsg });
-    if (callback) {
-      callback({ error: errorMsg });
-    }
-    return;
-  }
-
-  const { lobbyId } = data;
-
-  const lobby =
-    lobbyManager.getLobby(lobbyId) ?? (await refreshLobbyFromSupabase(lobbyId));
-  if (!lobby) {
-    const errorMsg = 'Lobby not found';
-    socket.emit('error', { message: errorMsg });
-    if (callback) {
-      callback({ error: errorMsg });
-    }
-    return;
-  }
-
-  if (playerId !== lobby.hostId) {
-    const errorMsg = 'Only the host can start the game';
-    socket.emit('error', { message: errorMsg });
-    if (callback) {
-      callback({ error: errorMsg });
-    }
-    return;
-  }
-
-  try {
-    const { gameId } = await startLobbyGameForHost(lobbyId, playerId);
-    if (callback) {
-      callback({ success: true, gameId });
-    }
-  } catch (error) {
-    const message =
-      error instanceof Error ? error.message : 'Failed to start game';
-    logger.error('Error starting lobby game:', message);
-    if (callback) {
-      callback({ error: message });
-    }
-    socket.emit('error', { message });
-  }
-}
+// Game lifecycle functions moved to GameManager
 
 // Lobby handlers moved to LobbyManager
 
@@ -1294,7 +991,7 @@ async function handleLogin(socket: Socket, data: LoginData): Promise<void> {
     socket.emit('loggedIn', loggedInData);
     emitConnectedPlayers();
 
-    sendMostRecentGameData(socket);
+    gameManager.sendMostRecentGameData(socket);
   } catch (error) {
     const message = error instanceof Error ? error.message : 'Login failed';
     logger.error('Supabase login failed:', message);
@@ -1317,502 +1014,6 @@ function emitToLobbyPlayers(
   payload?: unknown
 ): void {
   io.to(lobbyId).emit(event, payload);
-}
-
-async function handleRestartGame(socket: Socket): Promise<void> {
-  const playerId = connectionManager.getPlayerId(socket.id);
-  if (!playerId) {
-    logger.error('Player ID not found for socket:', socket.id);
-    socket.emit('error', { message: 'Not logged in' });
-    return;
-  }
-
-  // Get the lobby this player is in
-  const lobbyId = lobbyManager.getLobbyIdForPlayer(playerId);
-  if (!lobbyId) {
-    logger.error('Player not in a lobby:', playerId);
-    socket.emit('error', { message: 'Not in a lobby' });
-    return;
-  }
-
-  const lobby = lobbyManager.getLobby(lobbyId);
-  if (!lobby) {
-    logger.error('Lobby not found:', lobbyId);
-    socket.emit('error', { message: 'Lobby not found' });
-    return;
-  }
-
-  // Only allow the host to restart the game
-  if (playerId !== lobby.hostId) {
-    logger.error(
-      'Cannot restart game - player is not the host:',
-      playerId,
-      'host:',
-      lobby.hostId
-    );
-    socket.emit('error', { message: 'Only the host can restart the game' });
-    return;
-  }
-
-  // Check if the game is finished (not in progress)
-  if (lobby.status !== 'in_progress' && lobby.status !== 'finished') {
-    logger.error('Cannot restart game - lobby is not in progress:', lobbyId);
-    socket.emit('error', { message: 'Game is not in progress' });
-    return;
-  }
-
-  logger.info(`Restarting game for lobby: ${lobby.name} (${lobbyId})`);
-
-  // Cancel current game loop and clear artifacts
-  clearActiveGameArtifacts(lobbyId);
-
-  // Emit gameReset to clear client state
-  emitToLobbyPlayers(lobby.id, 'gameReset');
-
-  // Immediately start a new game with the same lobby/players
-  // Reuse the logic from handleStartLobbyGame but without the waiting status check
-  // Clean up any existing bots before creating new ones
-  cleanupBots(lobby.id);
-
-  // Build playersInfo from lobby members (humans only, no bots yet)
-  const playersInfo: PlayerIdAndName[] = lobby.players.map(p => ({
-    id: p.playerId,
-    name: p.displayName,
-  }));
-
-  // Calculate bots needed
-  const targetCount =
-    lobby.maxPlayers ?? lobby.playerCount ?? playersInfo.length;
-  const botsNeeded = Math.max(0, targetCount - playersInfo.length);
-  logger.info(
-    `Restarting lobby game: ${lobby.name} with ${playersInfo.length} humans and ${botsNeeded} bots needed`
-  );
-
-  // Create bots
-  const botNames = ['Bot Alex', 'Bot Morgan', 'Bot Jordan'];
-  const newBotIds: string[] = [];
-  for (let i = 0; i < botsNeeded; i++) {
-    const botName = botNames[i] || `Bot ${i + 1}`;
-    const botAgent = new ExhaustiveSimpleAgent();
-    const botId = `${botAgent.playerId}-${Date.now()}-${i}`;
-    // Update the agent's playerId to match the generated botId
-    botAgent.playerId = botId;
-    playersInfo.push({ id: botId, name: botName });
-    const botPlayerInfo: PlayerInfo = {
-      id: botId,
-      name: botName,
-      agent: botAgent,
-    };
-    connectionManager.setPlayer(botId, botPlayerInfo);
-    newBotIds.push(botId);
-    logger.info(`Added bot: ${botName} (ID: ${botId})`);
-  }
-
-  // Create GameLoop using players from the lobby
-  const agents: Map<string, GameAgent> = new Map();
-  // Populate human agents
-  lobby.players.forEach(p => {
-    const info = connectionManager.getPlayer(p.playerId);
-    if (info) agents.set(info.id, info.agent);
-  });
-  // Populate bot agents
-  newBotIds.forEach(id => {
-    const info = connectionManager.getPlayer(id);
-    if (info) agents.set(info.id, info.agent);
-  });
-
-  // Store bot IDs for cleanup after game ends
-  currentGameBotIdsByLobbyId.set(lobby.id, newBotIds);
-
-  // Filter out disconnected players - only include players who have agents
-  const validPlayersInfo = playersInfo.filter(p => agents.has(p.id));
-  if (validPlayersInfo.length !== playersInfo.length) {
-    const disconnectedPlayers = playersInfo.filter(p => !agents.has(p.id));
-    logger.warn(
-      `[handleRestartGame] Filtering out ${
-        disconnectedPlayers.length
-      } disconnected players: ${disconnectedPlayers
-        .map(p => p.name)
-        .join(', ')}`
-    );
-  }
-  if (validPlayersInfo.length < 2) {
-    throw new Error('Not enough connected players to restart game');
-  }
-
-    const gameLoop = new GameLoop(validPlayersInfo);
-    agents.forEach((agent, id) => gameLoop.addAgent(id, agent));
-    gameLoopsByLobbyId.set(lobby.id, gameLoop);
-    currentRoundGameEventsByLobbyId.set(lobby.id, []);
-    await persistenceService.createSupabaseGameForLobby(
-      lobby,
-      validPlayersInfo,
-      gameLoop,
-      supabaseGameIdByLobbyId,
-      SUPABASE_AUTH_ENABLED
-    );
-
-  // Set up gameSnapshot listener to send redacted snapshots to all clients
-  let firstSnapshotEmitted = false;
-  gameLoop.on('gameSnapshot', (newSnapshot: GameSnapshot) => {
-    mostRecentGameSnapshotByLobbyId.set(lobby.id, newSnapshot);
-    const existingEvents = currentRoundGameEventsByLobbyId.get(lobby.id) || [];
-    const updatedEvents = [...existingEvents, newSnapshot.gameEvent];
-    const isStartRound =
-      newSnapshot.gameEvent.actionType === ActionType.START_ROUND;
-    const roundEvents = isStartRound ? [newSnapshot.gameEvent] : updatedEvents;
-    if (isStartRound) {
-      logger.debug(
-        `[Supabase] START_ROUND detected, resetting event collection for lobby ${lobby.id}`
-      );
-      roundStartSnapshotByLobbyId.set(lobby.id, newSnapshot);
-      currentRoundGameEventsByLobbyId.set(lobby.id, roundEvents);
-    } else {
-      currentRoundGameEventsByLobbyId.set(lobby.id, updatedEvents);
-      logger.debug(
-        `[Supabase] Collected event ${newSnapshot.gameEvent.actionType} (lobby ${lobby.id}, total events: ${updatedEvents.length})`
-      );
-    }
-
-    if (
-      newSnapshot.gameEvent.actionType === ActionType.READY_FOR_NEXT_ROUND ||
-      newSnapshot.gameEvent.actionType === ActionType.WIN
-    ) {
-      const eventsToPersist =
-        currentRoundGameEventsByLobbyId.get(lobby.id) || [];
-      logger.info(
-        `[Supabase] Triggering persistence for ${newSnapshot.gameEvent.actionType} (lobby ${lobby.id}, ${eventsToPersist.length} events collected)`
-      );
-      // Use void to fire-and-forget, but ensure errors are logged in persistRoundHistory
-      void persistenceService
-        .persistRoundHistory(
-          lobby.id,
-          newSnapshot,
-          supabaseGameIdByLobbyId,
-          currentRoundGameEventsByLobbyId,
-          roundStartSnapshotByLobbyId,
-          SUPABASE_AUTH_ENABLED
-        )
-        .catch(error => {
-          logger.error(
-            `[Supabase] Unhandled error in persistRoundHistory for lobby ${lobby.id}`,
-            error
-          );
-        });
-    }
-
-    // Send redacted snapshots to all players
-    sendRedactedSnapshotToAllPlayers(
-      gameLoop,
-      newSnapshot,
-      roundEvents,
-      lobby.id
-    );
-
-    // After the first snapshot, ensure all clients received it by re-emitting
-    if (!firstSnapshotEmitted) {
-      firstSnapshotEmitted = true;
-      setTimeout(() => {
-        const latestSnapshot = mostRecentGameSnapshotByLobbyId.get(lobby.id);
-        const latestEvents =
-          currentRoundGameEventsByLobbyId.get(lobby.id) || [];
-        if (latestSnapshot === newSnapshot) {
-          logger.info(
-            'Re-emitting first game snapshot to ensure all clients received it'
-          );
-          sendRedactedSnapshotToAllPlayers(
-            gameLoop,
-            latestSnapshot,
-            latestEvents,
-            lobby.id
-          );
-        }
-      }, 100);
-    }
-  });
-
-  // Map lobby -> game
-  const gameId = gameLoop.cribbageGame.getGameState().id;
-  gameIdByLobbyId.set(lobby.id, gameId);
-
-  // Update lobby status (keep as in_progress, don't reset to waiting)
-  lobby.status = 'in_progress';
-  lobby.finishedAt = null;
-  lobby.disconnectedPlayerIds = [];
-  io.emit('lobbyUpdated', lobby);
-
-  // Notify lobby members of the game restart (same as game start)
-  // Small delay to ensure clients have processed gameReset
-  setTimeout(() => {
-    io.emit('gameStartedFromLobby', {
-      lobbyId: lobby.id,
-      gameId,
-      players: validPlayersInfo,
-    });
-  }, 50);
-
-  // Start the game loop (this will emit snapshots as the game progresses)
-  startGame(lobby.id).catch(error => {
-    logger.error('[handleRestartGame] Error in game loop:', error);
-  });
-
-  logger.info(`Game restarted. New game started for lobby ${lobby.name}.`);
-}
-
-/**
- * Send redacted game snapshot and events to all players in a game
- */
-function sendRedactedSnapshotToAllPlayers(
-  gameLoop: GameLoop,
-  snapshot: GameSnapshot,
-  roundEvents: GameEvent[],
-  lobbyId: string
-): void {
-  const gameState = gameLoop.cribbageGame.getGameState();
-
-  // Send redacted snapshot to each player
-  gameState.players.forEach(player => {
-    const socketId = connectionManager.getSocketId(player.id);
-    if (!socketId) {
-      // Player might be a bot or disconnected - skip
-      return;
-    }
-
-    const socket = io.sockets.sockets.get(socketId);
-    if (!socket) {
-      // Socket not found - skip
-      return;
-    }
-
-    // Update WebSocketAgent with the latest snapshot if this is a human player
-    const playerInfo = connectionManager.getPlayer(player.id);
-    if (playerInfo && playerInfo.agent instanceof WebSocketAgent) {
-      playerInfo.agent.updateGameSnapshot(snapshot);
-    }
-
-    // Get redacted state and event for this player
-    const redactedGameState = gameLoop.cribbageGame.getRedactedGameState(
-      player.id
-    );
-    const redactedGameEvent = gameLoop.cribbageGame.getRedactedGameEvent(
-      snapshot.gameEvent,
-      player.id
-    );
-    const redactedSnapshot: GameSnapshot = {
-      gameState: redactedGameState,
-      gameEvent: redactedGameEvent,
-      pendingDecisionRequests: snapshot.pendingDecisionRequests,
-    };
-
-    // Send redacted snapshot
-    socket.emit('gameSnapshot', redactedSnapshot);
-
-    // Send redacted round events
-    const redactedRoundEvents = roundEvents.map(event =>
-      gameLoop.cribbageGame.getRedactedGameEvent(event, player.id)
-    );
-    socket.emit('currentRoundGameEvents', redactedRoundEvents);
-  });
-}
-
-function sendMostRecentGameData(socket: Socket): void {
-  logger.info('Sending most recent game data to client');
-
-  // Find which player this socket belongs to
-  const playerId = connectionManager.getPlayerId(socket.id);
-
-  if (!playerId) {
-    logger.error('Could not find player ID for socket:', socket.id);
-    return;
-  }
-
-  const lobbyId = lobbyManager.getLobbyIdForPlayer(playerId);
-  if (!lobbyId) {
-    logger.warn(
-      `Player ${playerId} is not in a lobby; skipping game state send.`
-    );
-    socket.emit('currentRoundGameEvents', []);
-    return;
-  }
-
-  const activeGameLoop = gameLoopsByLobbyId.get(lobbyId);
-  const mostRecentGameSnapshot = mostRecentGameSnapshotByLobbyId.get(lobbyId);
-  const roundEvents = currentRoundGameEventsByLobbyId.get(lobbyId) || [];
-
-  if (!activeGameLoop || !mostRecentGameSnapshot) {
-    logger.warn(
-      `No active game loop or snapshot for lobby ${lobbyId} when attempting to send game data`
-    );
-    socket.emit('currentRoundGameEvents', []);
-    return;
-  }
-
-  // Check if player exists in the game before trying to get redacted state
-  const currentGameState = activeGameLoop.cribbageGame.getGameState();
-  const playerExistsInGame = currentGameState.players.some(
-    p => p.id === playerId
-  );
-
-  if (!playerExistsInGame) {
-    logger.info(
-      `Player ${playerId} not found in game for lobby ${lobbyId}. Skipping game state send.`
-    );
-    // Still send empty arrays for consistency
-    socket.emit('currentRoundGameEvents', []);
-    return;
-  }
-
-  // Update WebSocketAgent with the latest snapshot
-  const playerInfo = connectionManager.getPlayer(playerId);
-  if (playerInfo && playerInfo.agent instanceof WebSocketAgent) {
-    playerInfo.agent.updateGameSnapshot(mostRecentGameSnapshot);
-  }
-
-  const redactedGameState =
-    activeGameLoop.cribbageGame.getRedactedGameState(playerId);
-  const redactedGameEvent = activeGameLoop.cribbageGame.getRedactedGameEvent(
-    mostRecentGameSnapshot.gameEvent,
-    playerId
-  );
-  const redactedSnapshot: GameSnapshot = {
-    gameState: redactedGameState,
-    gameEvent: redactedGameEvent,
-    pendingDecisionRequests: mostRecentGameSnapshot.pendingDecisionRequests, // Include pending requests
-  };
-  socket.emit('gameSnapshot', redactedSnapshot);
-
-  const redactedRoundEvents = roundEvents.map(event =>
-    activeGameLoop.cribbageGame.getRedactedGameEvent(event, playerId)
-  );
-  socket.emit('currentRoundGameEvents', redactedRoundEvents);
-}
-
-async function startGame(lobbyId: string): Promise<void> {
-  const gameLoop = gameLoopsByLobbyId.get(lobbyId);
-  if (!gameLoop) {
-    logger.error(
-      `[startGame()] Game loop not initialized for lobby ${lobbyId}. Cannot start game.`
-    );
-    return;
-  }
-
-  // Store reference to current game loop to check if it was cancelled or replaced
-  const currentGameLoop = gameLoop;
-
-  try {
-    logger.info('Starting game loop...');
-    const winner = await currentGameLoop.playGame();
-
-    // Check if this game loop was cancelled (replaced by a new one)
-    if (gameLoopsByLobbyId.get(lobbyId) !== currentGameLoop) {
-      logger.info('[startGame()] Game loop was replaced, ignoring completion');
-      return;
-    }
-
-    const supabaseGameId = supabaseGameIdByLobbyId.get(lobbyId);
-    if (supabaseGameId) {
-      const latestSnapshot = mostRecentGameSnapshotByLobbyId.get(lobbyId);
-      if (latestSnapshot) {
-        await persistenceService.persistRoundHistory(
-          lobbyId,
-          latestSnapshot,
-          supabaseGameIdByLobbyId,
-          currentRoundGameEventsByLobbyId,
-          roundStartSnapshotByLobbyId,
-          SUPABASE_AUTH_ENABLED
-        );
-      }
-      const finalState = currentGameLoop.cribbageGame.getGameState();
-      const finalScores = finalState.players.map(player => ({
-        playerId: toUuidOrNull(player.id),
-        playerName: player.name,
-        score: player.score,
-        isWinner: winner ? player.id === winner : false,
-      }));
-      await completeGameRecord({
-        gameId: supabaseGameId,
-        winnerId: toUuidOrNull(winner),
-        finalState,
-        finalScores,
-        roundCount: finalState.roundNumber,
-        endedAt: new Date(),
-      });
-    }
-
-    // Wait a brief moment to ensure the final snapshot with Phase.END is sent to all clients
-    // The endGame() call emits a gameSnapshot event which needs to be processed and sent
-    await new Promise(resolve => setTimeout(resolve, 100));
-
-    // Check again if game loop was replaced during the wait
-    if (gameLoopsByLobbyId.get(lobbyId) !== currentGameLoop) {
-      logger.info(
-        '[startGame()] Game loop was replaced during wait, ignoring completion'
-      );
-      return;
-    }
-
-    // Clear gameLoop after game ends so a new game can be started
-    // Only clear if this is still the current game loop (hasn't been replaced)
-    if (gameLoopsByLobbyId.get(lobbyId) === currentGameLoop) {
-      logger.info('Game ended. Clearing game loop to allow new game.');
-      currentGameLoop.removeAllListeners();
-      gameLoopsByLobbyId.delete(lobbyId);
-    } else {
-      logger.info(
-        '[startGame()] Game loop was replaced, not clearing (new game already started)'
-      );
-    }
-
-    // Clean up bots that were created for this game
-    cleanupBots(lobbyId);
-
-    gameIdByLobbyId.delete(lobbyId);
-    supabaseGameIdByLobbyId.delete(lobbyId);
-    roundStartSnapshotByLobbyId.delete(lobbyId);
-
-    const completedLobby = lobbyManager.getLobby(lobbyId);
-    if (completedLobby) {
-      completedLobby.status = 'finished';
-      completedLobby.finishedAt = Date.now();
-      completedLobby.disconnectedPlayerIds = [];
-      io.emit('lobbyUpdated', completedLobby);
-    }
-
-    if (SUPABASE_AUTH_ENABLED) {
-      try {
-        await getServiceClient()
-          .from('lobbies')
-          .update({ status: 'finished' })
-          .eq('id', lobbyId);
-      } catch (updateError) {
-        logger.error(
-          '[startGame] Failed to mark lobby finished in Supabase',
-          updateError
-        );
-      }
-    }
-
-    io.emit('gameOver', winner);
-  } catch (error) {
-    // If game loop was cancelled, that's expected - just log and return
-    if (error instanceof Error && error.message === 'Game loop was cancelled') {
-      logger.info('[startGame()] Game loop was cancelled, cleaning up');
-      // Clean up bots even if cancelled
-      if (gameLoopsByLobbyId.get(lobbyId) === currentGameLoop) {
-        cleanupBots(lobbyId);
-        gameLoopsByLobbyId.delete(lobbyId);
-      }
-      gameIdByLobbyId.delete(lobbyId);
-      supabaseGameIdByLobbyId.delete(lobbyId);
-      roundStartSnapshotByLobbyId.delete(lobbyId);
-      currentRoundGameEventsByLobbyId.delete(lobbyId);
-      mostRecentGameSnapshotByLobbyId.delete(lobbyId);
-      return;
-    }
-    // Otherwise, rethrow the error
-    logger.error('[startGame()] Error during game:', error);
-    throw error;
-  }
 }
 
 // Start the server
